Split scan_controller in utils and scan_reconciler
To make it easier to refactor the scan_controller we started by
splitting it into utils and scan_reconciler.
